fix: move themes to theme folder
